pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "RE: From Zero to Agentic!!"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.15",
    "fastapi>=0.135.1",
    "graphdatascience>=1.20",
    "ipython>=9.11.0",
    "langchain>=1.2.10",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.11",
    "langchain-text-splitters>=1.1.1",
    "llama-index>=0.14.15",
    "llama-index-embeddings-ollama>=0.8.6",
    "llama-index-llms-ollama>=0.9.1",
    "mineru-vl-utils>=0.1.22",
    "neo4j-graphrag>=1.0.0,<1.13.0",
    "numpy>=2.1.0",
    "pydantic-ai>=1.67.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.55.0",
    "transformers>=5.3.0",
]
